dev: Adds initial Docker Compose setup
Sets up Docker Compose environment for local development.

This commit introduces a `docker-compose.yml` file to define and manage multi-container Docker applications. It includes services for PHP, MariaDB, Redis, and Nginx, along with necessary configurations for networking, volumes, and dependencies.
This allows for easier setup and management of the Freshware application in a local environment.